requirements.txt
numpy==1.19.5
mujoco-py==1.50.1.68
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
tqdm==4.32.1
ray[rllib,debug]==0.6.4
gtimer==1.0.0b5
gym==0.12.0
pybullet==3.0.8
git+git://github.com/hartikainen/serializable.git
